// src/auth/model/permission.enum.ts
export enum Permission {
  CREATE_GAMES = 'create:games',
  READ_GAMES = 'read:games',
  UPDATE_GAMES = 'update:games',
  DELETE_GAMES = 'delete:games',
  CREATE_ROUNDS = 'create:rounds',
  READ_ROUNDS = 'read:rounds',
  UPDATE_ROUNDS = 'update:rounds',
  DELETE_ROUNDS = 'delete:rounds',
  CREATE_EVENTS = 'create:events',
  READ_EVENTS = 'read:events',
  UPDATE_EVENTS = 'update:events',
  DELETE_EVENTS = 'delete:events',
  CREATE_EVENT_TYPES = 'create:event_types',
  READ_EVENT_TYPES = 'read:event_types',
  UPDATE_EVENT_TYPES = 'update:event_types',
  DELETE_EVENT_TYPES = 'delete:event_types',
  CREATE_PLAYERS = 'create:players',
  READ_PLAYERS = 'read:players',
  UPDATE_PLAYERS = 'update:players',
  DELETE_PLAYERS = 'delete:players',
  READ_STATISTICS = 'read:statistics',
  READ_PAYMENTS = 'read:payments',
  UPDATE_PAYMENTS = 'update:payments',
  READ_DEBUGGING_INFO = 'read:debugging-info'
}